Uncorrupted by Capitalism
- spacing
- added pre hook to rollConcentrationSave

+++ scripts/_promptListeners.mjs
@@ -17,20 +17,20 @@ export function setHooks_promptListeners() {
 async function clickConcentrationPrompt(event) {
     // get the target of the mouse click.
     let button = event.target?.closest("button");
-    if (!button) button = event.target?.closest("[name='render-item-sheet']");
-    if (!button) return;
+    if ( !button ) button = event.target?.closest("[name='render-item-sheet']");
+    if ( !button ) return;
 
-    if (button.name === "saving-throw-button") {
+    if ( button.name === "saving-throw-button" ) {
         const { dc, saveType, actorUuid } = button.dataset;
         const actor = fromUuidSync(actorUuid);
         return actor.rollConcentrationSave(saveType, { targetValue: dc });
     }
-    else if (button.name === "delete-concentration-button") {
+    else if ( button.name === "delete-concentration-button" ) {
         const { effectUuid } = button.dataset;
         const effect = fromUuidSync(effectUuid);
         return deleteDialog(effect, event);
     }
-    else if (button.name === "render-item-sheet") {
+    else if ( button.name === "render-item-sheet" ) {
         const { itemUuid } = button.dataset;
         const item = fromUuidSync(itemUuid);
         return item.sheet.render(true);
@@ -38,7 +38,8 @@ async function clickConcentrationPrompt(event) {
 }
 
 async function deleteDialog(effect, event) {
-    if (event.shiftKey) {
+    if ( !effect ) return;
+    if ( event.shiftKey ) {
         return effect.delete();
     }
     const name = effect.getFlag(MODULE, "data.itemData.name");
